pyproject.toml
[tool.poetry]
name = "sbi-pdf2text"
version = "0.1.0"
description = ""
authors = ["Kei198403@gmail.com"]

[tool.poetry.dependencies]
python = "^3.9"
camelot-py = {extras = ["plot"], version = "^0.10.1"}
mojimoji = "^0.0.12"
ghostscript = "^0.7"
opencv-python = "^4.5.5"
pdftopng = "^0.2.3"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
mypy = "^0.942"
autopep8 = "^1.6.0"
pylint = "^2.13.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
